move to pkg/util for share with pkg/wasm
