+++ istioctl/pkg/writer/compare/sds/util.go
@@ -23,7 +23,7 @@ import (
 	envoy_admin "github.com/envoyproxy/go-control-plane/envoy/admin/v3"
 	auth "github.com/envoyproxy/go-control-plane/envoy/extensions/transport_sockets/tls/v3"
 
-	"istio.io/istio/istioctl/pkg/util/configdump"
+	"istio.io/istio/pkg/util/configdump"
 	"istio.io/pkg/log"
 )
